src/requirements.txt


kedro~=0.18.6  # python <3.11
kedro-viz
kedro-datasets[pandas.CSVDataSet]~=1.0.0
kedro-mlflow
pandas
geopandas
tqdm
unidecode
black
srai[all] @ git+https://github.com/srai-lab/srai.git@highway2vec
ipykernel
ipywidgets
mapclassify
folium
matplotlib
pytorch-lightning[extra]
tensorboard
torch-tb-profiler